reduced calculations

+++ src/Face.java
@@ -25,22 +25,22 @@ public class Face
     }
 
     public void draw(Graphics2D g2) {
-        poly = new Polygon(new int[] {(int)l1.getPointOne().get2Dx(), 
-                (int)l1.getPointTwo().get2Dx(), 
-                (int)l2.getPointOne().get2Dx(), 
-                (int)l2.getPointTwo().get2Dx(), 
-                (int)l3.getPointOne().get2Dx(), 
-                (int)l3.getPointTwo().get2Dx(), 
-                (int)l4.getPointTwo().get2Dx(), 
-                (int)l4.getPointOne().get2Dx(),},
-            new int[] {(int)l1.getPointOne().get2Dy(), 
-                (int)l1.getPointTwo().get2Dy(), 
-                (int)l2.getPointOne().get2Dy(), 
-                (int)l2.getPointTwo().get2Dy(), 
-                (int)l3.getPointOne().get2Dy(), 
-                (int)l3.getPointTwo().get2Dy(), 
-                (int)l4.getPointTwo().get2Dy(), 
-                (int)l4.getPointOne().get2Dy(),},
+        poly = new Polygon(new int[] {l1.getPointOne().get2Dx(), 
+                l1.getPointTwo().get2Dx(), 
+                l2.getPointOne().get2Dx(), 
+                l2.getPointTwo().get2Dx(), 
+                l3.getPointOne().get2Dx(), 
+                l3.getPointTwo().get2Dx(), 
+                l4.getPointTwo().get2Dx(), 
+                l4.getPointOne().get2Dx(),},
+            new int[] {l1.getPointOne().get2Dy(), 
+                l1.getPointTwo().get2Dy(), 
+                l2.getPointOne().get2Dy(), 
+                l2.getPointTwo().get2Dy(), 
+                l3.getPointOne().get2Dy(), 
+                l3.getPointTwo().get2Dy(), 
+                l4.getPointTwo().get2Dy(), 
+                l4.getPointOne().get2Dy(),},
             8);
 
         g2.setColor(color);
